requirements.txt
paramiko
openai
pyyaml
aiohttp
aiosqlite
fastapi
uvicorn[standard]
jinja2
python-multipart
